[inference] Consolidate WorkerWrap MoE reload regression into test_weight_sync.py
Move the legacy `WorkerWrap.load_weights` MoE reload regression added in
the prior commit out of its standalone `test_vllm_worker_moe_reload.py`
and into `test_weight_sync.py`, alongside the existing NCCL-broadcast and
CUDA-IPC weight-sync scenarios.
